Adding some tests for the publisher project
I needed to create a wrapper for the QueueClient as this was returning a concrete object making the rest of the code harder to test.
I have also added InternalsVisibleTo for two namespaces so we can test internal logic: one was for this repository testing library, and the other was for Mock so it can mock the interfaces.